ci: fail if standalone.py changed and standalone.tpl did not
This commit adds to the pre-commit hook a check to verify that changes
to standalone.py are accompanied by corresponding updates to
standalone.tpl. The check will fail if standalone.py is updated without
updating standalone.tpl ONLY if no components file(s) is/are updated.
This ensures consistency between standalone.py and standalone.tpl, and
prompts the user to run 'make standalone' if necessary. The check can be
bypassed by adding the 'ignore-standalone-change' label to the pull
request.

Signed-off-by: Sébastien Han <seb@redhat.com>